test: separate out buffer copy tests
