+++ test/parallel/test-buffer-alloc.js
@@ -4,9 +4,6 @@ var assert = require('assert');
 
 var Buffer = require('buffer').Buffer;
 
-// counter to ensure unique value is always copied
-var cntr = 0;
-
 var b = Buffer.allocUnsafe(1024);
 
 console.log('b.length == %d', b.length);
@@ -35,129 +32,6 @@ var e = Buffer.from(ui32);
 for (const [index, value] of e.entries()) {
   assert.strictEqual(value, ui32[index]);
 }
-
-{
-  // copy 512 bytes, from 0 to 512.
-  b.fill(++cntr);
-  c.fill(++cntr);
-  const copied = b.copy(c, 0, 0, 512);
-  console.log('copied %d bytes from b into c', copied);
-  assert.strictEqual(512, copied);
-  for (let i = 0; i < c.length; i++) {
-    assert.strictEqual(b[i], c[i]);
-  }
-}
-
-{
-  // copy c into b, without specifying sourceEnd
-  b.fill(++cntr);
-  c.fill(++cntr);
-  const copied = c.copy(b, 0, 0);
-  console.log('copied %d bytes from c into b w/o sourceEnd', copied);
-  assert.strictEqual(c.length, copied);
-  for (let i = 0; i < c.length; i++) {
-    assert.strictEqual(c[i], b[i]);
-  }
-}
-
-{
-  // copy c into b, without specifying sourceStart
-  b.fill(++cntr);
-  c.fill(++cntr);
-  const copied = c.copy(b, 0);
-  console.log('copied %d bytes from c into b w/o sourceStart', copied);
-  assert.strictEqual(c.length, copied);
-  for (let i = 0; i < c.length; i++) {
-    assert.strictEqual(c[i], b[i]);
-  }
-}
-
-{
-  // copy longer buffer b to shorter c without targetStart
-  b.fill(++cntr);
-  c.fill(++cntr);
-  const copied = b.copy(c);
-  console.log('copied %d bytes from b into c w/o targetStart', copied);
-  assert.strictEqual(c.length, copied);
-  for (let i = 0; i < c.length; i++) {
-    assert.strictEqual(b[i], c[i]);
-  }
-}
-
-{
-  // copy starting near end of b to c
-  b.fill(++cntr);
-  c.fill(++cntr);
-  const copied = b.copy(c, 0, b.length - Math.floor(c.length / 2));
-  console.log('copied %d bytes from end of b into beginning of c', copied);
-  assert.strictEqual(Math.floor(c.length / 2), copied);
-  for (let i = 0; i < Math.floor(c.length / 2); i++) {
-    assert.strictEqual(b[b.length - Math.floor(c.length / 2) + i], c[i]);
-  }
-  for (let i = Math.floor(c.length / 2) + 1; i < c.length; i++) {
-    assert.strictEqual(c[c.length - 1], c[i]);
-  }
-}
-
-{
-  // try to copy 513 bytes, and check we don't overrun c
-  b.fill(++cntr);
-  c.fill(++cntr);
-  const copied = b.copy(c, 0, 0, 513);
-  console.log('copied %d bytes from b trying to overrun c', copied);
-  assert.strictEqual(c.length, copied);
-  for (let i = 0; i < c.length; i++) {
-    assert.strictEqual(b[i], c[i]);
-  }
-}
-
-{
-  // copy 768 bytes from b into b
-  b.fill(++cntr);
-  b.fill(++cntr, 256);
-  const copied = b.copy(b, 0, 256, 1024);
-  console.log('copied %d bytes from b into b', copied);
-  assert.strictEqual(768, copied);
-  for (let i = 0; i < b.length; i++) {
-    assert.strictEqual(cntr, b[i]);
-  }
-}
-
-// copy string longer than buffer length (failure will segfault)
-var bb = Buffer.allocUnsafe(10);
-bb.fill('hello crazy world');
-
-
-// try to copy from before the beginning of b
-assert.doesNotThrow(() => { b.copy(c, 0, 100, 10); });
-
-// copy throws at negative sourceStart
-assert.throws(function() {
-  Buffer.allocUnsafe(5).copy(Buffer.allocUnsafe(5), 0, -1);
-}, RangeError);
-
-{
-  // check sourceEnd resets to targetEnd if former is greater than the latter
-  b.fill(++cntr);
-  c.fill(++cntr);
-  const copied = b.copy(c, 0, 0, 1025);
-  console.log('copied %d bytes from b into c', copied);
-  for (let i = 0; i < c.length; i++) {
-    assert.strictEqual(b[i], c[i]);
-  }
-}
-
-// throw with negative sourceEnd
-console.log('test copy at negative sourceEnd');
-assert.throws(function() {
-  b.copy(c, 0, 0, -1);
-}, RangeError);
-
-// when sourceStart is greater than sourceEnd, zero copied
-assert.equal(b.copy(c, 0, 100, 10), 0);
-
-// when targetStart > targetLength, zero copied
-assert.equal(b.copy(c, 512, 0, 10), 0);
 
 var caught_error;
